Merged PR 114086: MANUAL CHANGE - Remove deprecated DataChanged property
Remove the deprecated DataChanged property.
This property is used by VB6's internal binding mechanism.
In JetNet, it was used to determine if the value of the checkboxes had changed and update the recordset when necessary. But this is not required since it is validated again in the inner conditions.

Related work items: #761579